# Connection\Manager: is_connected: Memoize