--- a/projects/packages/connection/changelog/update-is-connected
+++ b/projects/packages/connection/changelog/update-is-connected
@@ -1,0 +1,5 @@
+Significance: patch
+Type: changed
+Comment: Internal performance change for is_connected()
+
+