--- a/projects/packages/connection/tests/php/test-rest-endpoints.php
+++ b/projects/packages/connection/tests/php/test-rest-endpoints.php
@@ -110,6 +110,7 @@ class Test_REST_Endpoints extends TestCase {
 		Constants::$set_constants['JETPACK__API_BASE'] = 'https://jetpack.wordpress.com/jetpack.';
 
 		set_transient( 'jetpack_assumed_site_creation_date', '2020-02-28 01:13:27' );
+		$this->reset_connection_status();
 	}
 
 	/**
@@ -137,6 +138,20 @@ class Test_REST_Endpoints extends TestCase {
 		$_GET = array();
 
 		Connection_Rest_Authentication::init()->reset_saved_auth_state();
+		$this->reset_connection_status();
+	}
+
+	/**
+	 * Reset the connection status.
+	 * Needed because the connection status is memoized and not reset between tests.
+	 * WorDBless does not fire the options update hooks that would reset the connection status.
+	 */
+	public function reset_connection_status() {
+		static $manager = null;
+		if ( ! $manager ) {
+			$manager = new \Automattic\Jetpack\Connection\Manager();
+		}
+		$manager->reset_connection_status();
 	}
 
 	/**
@@ -1764,6 +1779,7 @@ class Test_REST_Endpoints extends TestCase {
 		}
 
 		add_filter( 'jetpack_options', array( $this, 'mock_jetpack_options' ), 10, 2 );
+		$this->reset_connection_status();
 	}
 
 	/**
@@ -1816,5 +1832,6 @@ class Test_REST_Endpoints extends TestCase {
 		}
 
 		remove_filter( 'jetpack_options', array( $this, 'mock_jetpack_options' ), 10 );
+		$this->reset_connection_status();
 	}
 }