--- a/projects/packages/connection/tests/php/test_package_version_tracker.php
+++ b/projects/packages/connection/tests/php/test_package_version_tracker.php
@@ -48,6 +48,7 @@ class Test_Package_Version_Tracker extends TestCase {
 	public function set_up() {
 		Constants::set_constant( 'JETPACK__WPCOM_JSON_API_BASE', 'https://public-api.wordpress.com' );
 		Sync_Settings::update_settings( array( 'disable' => true ) );
+		$this->reset_connection_status();
 	}
 
 	/**
@@ -66,6 +67,20 @@ class Test_Package_Version_Tracker extends TestCase {
 
 		delete_transient( Package_Version_Tracker::CACHED_FAILED_REQUEST_KEY );
 		delete_transient( Package_Version_Tracker::RATE_LIMITER_KEY );
+		$this->reset_connection_status();
+	}
+
+	/**
+	 * Reset the connection status.
+	 * Needed because the connection status is memoized and not reset between tests.
+	 * WorDBless does not fire the options update hooks that would reset the connection status.
+	 */
+	public function reset_connection_status() {
+		static $manager = null;
+		if ( ! $manager ) {
+			$manager = new \Automattic\Jetpack\Connection\Manager();
+		}
+		$manager->reset_connection_status();
 	}
 
 	/**
@@ -305,6 +320,7 @@ class Test_Package_Version_Tracker extends TestCase {
 	public function test_maybe_update_package_versions_with_sync_disabled_remote_request_success() {
 		\Jetpack_Options::update_option( 'blog_token', 'asdasd.123123' );
 		\Jetpack_Options::update_option( 'id', 1234 );
+		$this->reset_connection_status();
 
 		add_filter( 'pre_http_request', array( $this, 'intercept_http_request_success' ) );
 
@@ -360,6 +376,7 @@ class Test_Package_Version_Tracker extends TestCase {
 	public function test_maybe_update_package_versions_with_sync_disabled_remote_request_failure() {
 		\Jetpack_Options::update_option( 'blog_token', 'asdasd.123123' );
 		\Jetpack_Options::update_option( 'id', 1234 );
+		$this->reset_connection_status();
 
 		add_filter( 'pre_http_request', array( $this, 'intercept_http_request_failure' ) );
 
@@ -393,6 +410,7 @@ class Test_Package_Version_Tracker extends TestCase {
 
 		\Jetpack_Options::update_option( 'blog_token', 'asdasd.123123' );
 		\Jetpack_Options::update_option( 'id', 1234 );
+		$this->reset_connection_status();
 
 		add_filter( 'pre_http_request', array( $this, 'intercept_http_request_failure' ) );