--- a/projects/packages/connection/tests/php/test_plugin_storage.php
+++ b/projects/packages/connection/tests/php/test_plugin_storage.php
@@ -35,6 +35,7 @@ class Test_Plugin_Storage extends TestCase {
 	public function set_up() {
 		Constants::set_constant( 'JETPACK__WPCOM_JSON_API_BASE', 'https://public-api.wordpress.com' );
 		Sync_Settings::update_settings( array( 'disable' => true ) );
+		$this->reset_connection_status();
 	}
 
 	/**
@@ -61,6 +62,20 @@ class Test_Plugin_Storage extends TestCase {
 			$plugins->setAccessible( true );
 			$plugins->setValue( array() );
 		}
+		$this->reset_connection_status();
+	}
+
+	/**
+	 * Reset the connection status.
+	 * Needed because the connection status is memoized and not reset between tests.
+	 * WorDBless does not fire the options update hooks that would reset the connection status.
+	 */
+	public function reset_connection_status() {
+		static $manager = null;
+		if ( ! $manager ) {
+			$manager = new \Automattic\Jetpack\Connection\Manager();
+		}
+		$manager->reset_connection_status();
 	}
 
 	/**
@@ -71,6 +86,7 @@ class Test_Plugin_Storage extends TestCase {
 	public function test_update_active_plugins_option_without_sync_will_trigger_fallback() {
 		\Jetpack_Options::update_option( 'blog_token', 'asdasd.123123' );
 		\Jetpack_Options::update_option( 'id', 1234 );
+		$this->reset_connection_status();
 
 		add_filter( 'pre_http_request', array( $this, 'intercept_remote_request' ), 10, 3 );
 		Plugin_Storage::update_active_plugins_option();
@@ -109,6 +125,7 @@ class Test_Plugin_Storage extends TestCase {
 	public function test_maybe_update_active_connected_plugins_not_configured() {
 		\Jetpack_Options::update_option( 'blog_token', 'asdasd.123123' );
 		\Jetpack_Options::update_option( 'id', 1234 );
+		$this->reset_connection_status();
 
 		Plugin_Storage::upsert( 'dummy-slug' );
 		set_transient( Plugin_Storage::ACTIVE_PLUGINS_REFRESH_FLAG, microtime() );
@@ -130,6 +147,7 @@ class Test_Plugin_Storage extends TestCase {
 	public function test_maybe_update_active_connected_plugins_flag_not_set() {
 		\Jetpack_Options::update_option( 'blog_token', 'asdasd.123123' );
 		\Jetpack_Options::update_option( 'id', 1234 );
+		$this->reset_connection_status();
 
 		Plugin_Storage::upsert( 'dummy-slug' );
 		Plugin_Storage::configure();
@@ -151,6 +169,7 @@ class Test_Plugin_Storage extends TestCase {
 	public function test_maybe_update_active_connected_plugins_non_post_request() {
 		\Jetpack_Options::update_option( 'blog_token', 'asdasd.123123' );
 		\Jetpack_Options::update_option( 'id', 1234 );
+		$this->reset_connection_status();
 
 		Plugin_Storage::upsert( 'dummy-slug' );
 		Plugin_Storage::configure();
@@ -172,6 +191,7 @@ class Test_Plugin_Storage extends TestCase {
 	public function test_maybe_update_active_connected_plugins_success() {
 		\Jetpack_Options::update_option( 'blog_token', 'asdasd.123123' );
 		\Jetpack_Options::update_option( 'id', 1234 );
+		$this->reset_connection_status();
 
 		Plugin_Storage::upsert( 'dummy-slug' );
 		Plugin_Storage::configure();
@@ -195,6 +215,7 @@ class Test_Plugin_Storage extends TestCase {
 	public function test_maybe_update_active_connected_plugins_success_same_plugins() {
 		\Jetpack_Options::update_option( 'blog_token', 'asdasd.123123' );
 		\Jetpack_Options::update_option( 'id', 1234 );
+		$this->reset_connection_status();
 		$stored_value = array( 'dummy-slug' => array() );
 		update_option( Plugin_Storage::ACTIVE_PLUGINS_OPTION_NAME, $stored_value );
 
@@ -220,6 +241,7 @@ class Test_Plugin_Storage extends TestCase {
 	public function test_maybe_update_active_connected_plugins_success_same_count_different_plugins() {
 		\Jetpack_Options::update_option( 'blog_token', 'asdasd.123123' );
 		\Jetpack_Options::update_option( 'id', 1234 );
+		$this->reset_connection_status();
 		update_option( Plugin_Storage::ACTIVE_PLUGINS_OPTION_NAME, array( 'dummy-slug2' => array() ) );
 
 		Plugin_Storage::upsert( 'dummy-slug' );