--- a/projects/packages/protect-status/changelog/update-is-connected
+++ b/projects/packages/protect-status/changelog/update-is-connected
@@ -1,0 +1,5 @@
+Significance: patch
+Type: changed
+Comment: Internal unit test changes
+
+