--- a/projects/packages/stats/changelog/update-is-connected
+++ b/projects/packages/stats/changelog/update-is-connected
@@ -1,0 +1,5 @@
+Significance: patch
+Type: changed
+Comment: Internal unit test changes
+
+