## Proposed changes:

* Memoize the `is_connected()` function, so it's only figured out once per request
  * Also a short boolean optimization, for `a && b`, if `a` is false, we don't need to figure out `b`.

### Reasoning

I noticed it popping up a lot while I reviewed timelines for my local site running JP. I used `ac_start()` and `ac_stop()` to measure its timing ([link](https://github.com/mreishus/wp-misc/blob/master/debug-helpers.php#L90-L125)) before and after adding memoization.

**Before**
```
[             ] https://myJTsite.example.com/wp-cron.php?doing_wp_cron=1726077077.1958370208740234375000
[             a19432] AC Debug: 'is_connected1' ran 38 times, total time: 5.82 ms, avg time: 0.15 ms
[             ] https://myJTsite.example.com/
[             c43038] AC Debug: 'is_connected1' ran 32 times, total time: 4.33 ms, avg time: 0.14 ms
[             ] https://myJTsite.example.com/wp-admin/admin.php?page=stats&noheader&proxy&chart=admin-bar-hours-scale
[             adcb58] AC Debug: 'is_connected1' ran 45 times, total time: 3.83 ms, avg time: 0.09 ms
[             ] https://myJTsite.example.com/
[             d41c05] AC Debug: 'is_connected1' ran 31 times, total time: 3.8 ms, avg time: 0.12 ms
[             ] https://myJTsite.example.com/wp-admin/admin.php?page=stats&noheader&proxy&chart=admin-bar-hours-scale
[             1b0db2] AC Debug: 'is_connected1' ran 45 times, total time: 5.83 ms, avg time: 0.13 ms
[             ] https://myJTsite.example.com/
```

**after**
```
[             f07557] AC Debug: 'is_connected2' ran 31 times, total time: 0.17 ms, avg time: 0.01 ms
[             ] https://myJTsite.example.com/wp-admin/admin.php?page=stats&noheader&proxy&chart=admin-bar-hours-scale
[             baec98] AC Debug: 'is_connected2' ran 45 times, total time: 0.19 ms, avg time: 0 ms
[             ] https://myJTsite.example.com/
[             cdd611] AC Debug: 'is_connected2' ran 31 times, total time: 0.15 ms, avg time: 0 ms
[             ] https://myJTsite.example.com/wp-admin/admin.php?page=stats&noheader&proxy&chart=admin-bar-hours-scale
[             6dcbd5] AC Debug: 'is_connected2' ran 45 times, total time: 0.14 ms, avg time: 0 ms
[             ] https://myJTsite.example.com/
[             38bb3d] AC Debug: 'is_connected2' ran 31 times, total time: 0.11 ms, avg time: 0 ms
[             ] https://myJTsite.example.com/wp-admin/admin.php?page=stats&noheader&proxy&chart=admin-bar-hours-scale
[             299baf] AC Debug: 'is_connected2' ran 45 times, total time: 0.13 ms, avg time: 0 ms
[             ] https://myJTsite.example.com/wp-cron.php?doing_wp_cron=1726077292.4192719459533691406250
[             1f2e26] AC Debug: 'is_connected2' ran 36 times, total time: 0.16 ms, avg time: 0 ms
[             ] https://myJTsite.example.com/
[             50359e] AC Debug: 'is_connected2' ran 32 times, total time: 0.21 ms, avg time: 0.01 ms
[             ] https://myJTsite.example.com/wp-admin/admin.php?page=stats&noheader&proxy&chart=admin-bar-hours-scale
[             863e0f] AC Debug: 'is_connected2' ran 45 times, total time: 0.18 ms, avg time: 0 ms
```
### Concerns

- Would there be any need to invalidate this? For example, a request that starts out connected and then is later not connected? Or vice versa? I don't know the JP codebase well enough to know if this is something that would be accounted for. If so, it's not a big deal, I just don't know where that would be or how to test it.


### Other information:

- [ ] Have you written new tests for your changes, if applicable?
- [ ] Have you checked the E2E test CI results, and verified that your changes do not break them?
- [ ] Have you tested your changes on WordPress.com, if applicable (if so, you'll see a generated comment below with a script to run)?

## Jetpack product discussion
<!-- If you're an Automattician, include a shortlink to the p2 discussion with Jetpack Product here. -->
<!-- Make sure any changes to existing products have been discussed and agreed upon -->

## Does this pull request change what data or activity we track or use?
<!--- If so, please add the "[Status] Needs Privacy Updates" label and explain what changes there are. -->
<!--- Check existing Jetpack support documents for a preview of the information we need. -->

## Testing instructions:

* Unknown :( General connection working? Connect / Disconnect?