@oskosk 2024-09-11
@sergeymitr is also looking at this. He mentioned the checks are failing cause we're not resetting the state between tests /

I tested launching a site from scratch ([JN](https://jurassic.ninja/create?jetpack-beta&branches.jetpack=update/is-connected&wp-debug-log)), disconnecting, and reconnecting and it looked fine.