@sergeymitr 2024-09-11
I believe we should also reset self::$is_connected on [site registration](https://github.com/Automattic/jetpack/blob/3bef458c3bb34197e1628f4b3ac1fc04e810c5c6/projects/packages/connection/src/class-manager.php#L1244) and [disconnect](https://github.com/Automattic/jetpack/blob/3bef458c3bb34197e1628f4b3ac1fc04e810c5c6/projects/packages/connection/src/class-manager.php#L2172) when we know for sure that the value has changed.