@mreishus 2024-09-12
Ok, got the tests to pass at least. I want to consider these things
* Is there still a benefit after the invalidation added
* Possibly adding the hook(s) I need to the jetpack options class instead of invalidating on every write to the jetpack private options
* Seeing if there's a better way to deal with WorDBless